+++ src/main/java/com/fg/fruitstore/strategy/NoDiscountStrategy.java
@@ -1,9 +1,10 @@
 package com.fg.fruitstore.strategy;
+import java.math.BigDecimal;
 
 public class NoDiscountStrategy implements DiscountStrategy {
 
     @Override
-    public double applyDiscount(double originalTotal) {
+    public BigDecimal applyDiscount(BigDecimal originalTotal) {
         return originalTotal;
     }
 }
